refactor(evals): inline inference runner into command handler
