+++ lalamo/evals/__init__.py
@@ -1,6 +1,6 @@
 from lalamo.evals.cli import eval_app
 from lalamo.evals.datasets import REPO_TO_EVAL, EvalSpec, convert_dataset
-from lalamo.evals.inference import InferenceEngine, LalamoInferenceEngine, run_batch_generation, run_inference
+from lalamo.evals.inference import InferenceEngine, LalamoInferenceEngine
 
 __all__ = [
     "REPO_TO_EVAL",
@@ -9,6 +9,4 @@ __all__ = [
     "LalamoInferenceEngine",
     "convert_dataset",
     "eval_app",
-    "run_batch_generation",
-    "run_inference",
 ]
