+++ lalamo/evals/inference/__init__.py
@@ -1,12 +1,9 @@
 from lalamo.evals.inference.callbacks import BaseRunInferenceCallbacks, ConsoleRunInferenceCallbacks
 from lalamo.evals.inference.engines import InferenceEngine, LalamoInferenceEngine
-from lalamo.evals.inference.runner import run_batch_generation, run_inference
 
 __all__ = [
     "BaseRunInferenceCallbacks",
     "ConsoleRunInferenceCallbacks",
     "InferenceEngine",
     "LalamoInferenceEngine",
-    "run_batch_generation",
-    "run_inference",
 ]
